Bumping version to 0.1.5.
Version 0.1.4 was published and released with outdated documentation.
Documentation was updated later in
https://github.com/omni-network/omni-std/pull/18.
This commit bumps the version number so that a new release can be
published with updated docs.

This also resolves discrepency between github and npm pacakage releases.
Npm has some 0.1.5-alpha.x pre relases that are not represented on github.
We should introduce some "npm publish on github release" flow so that
the two do not diverge like this in the future.